evdev: Associate touchscreen devices with output up front
This isn't going to change over time, so just tracking it in the
evdev device is a little easier.  Also, we need to adjust for the
output position when transforming the device events to screen space.

(cherry picked from commit e7b5b41e9378791459d1e5b5016a6868ddf49ae1)

Signed-off-by: Tiago Vignatti <tiago.vignatti@intel.com>